[REFACTOR] Move shared components into components folder
