+++ src/client/react/stages/Registration/Home.js
@@ -3,8 +3,8 @@ import { Link } from 'react-router-dom';
 import { Button, Intent } from '@blueprintjs/core';
 import { connect } from 'react-redux';
 import ScrollAnimation from 'react-animate-on-scroll';
-import Title from '../Title';
-import Description from '../Description';
+import Title from '../components/Title';
+import Description from '../components/Description';
 
 
 const mapStateToProps = (state, ownProps) => {
