+++ src/client/react/stages/Registration/index.js
@@ -2,8 +2,8 @@ import React from 'react';
 import { BrowserRouter, Switch, Route, Redirect } from 'react-router-dom';
 import { NotFoundPage } from '../../pages';
 import AppContainer from '../../../../../lib/react/components/AppContainer';
-import Login from '../Login';
-import Pay from '../Pay';
+import Login from '../components/Login';
+import Pay from '../components/Pay';
 import Home from './Home';
 import Register from './Register';
 import Dashboard from './Dashboard';
